feature(frontend): Enabling Test Run Panels
