+++ web/package.json
@@ -49,9 +49,10 @@
     "react-flow-renderer": "10.3.7",
     "react-joyride": "2.5.3",
     "react-redux": "7.2.6",
-    "react-reflex": "4.0.9",
+    "react-reflex": "4.1.0",
     "react-router-dom": "6.2.1",
     "react-scripts": "5.0.1",
+    "react-spaces": "0.3.8",
     "react-syntax-highlighter": "15.5.0",
     "react-use": "17.4.0",
     "redux-first-history": "5.0.12",
