+++ web/src/components/ResizablePanels/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line no-restricted-exports
-export {default} from './Drawer';
+export {default} from './ResizablePanels';
