+++ web/src/components/RunDetailTrace/RunDetailTrace.styled.ts
@@ -1,6 +1,6 @@
 import {CloseCircleFilled} from '@ant-design/icons';
 import {Typography} from 'antd';
-import styled from 'styled-components';
+import styled, {css} from 'styled-components';
 
 export const Container = styled.div`
   display: flex;
@@ -62,4 +62,26 @@ export const NoMatchesContainer = styled.div`
 
 export const NoMatchesText = styled(Typography.Text)`
   color: ${({theme}) => theme.color.textSecondary};
+`;
+
+export const PanelContainer = styled.div<{$isOpen: boolean}>`
+  background-color: ${({theme}) => theme.color.white};
+  box-shadow: 0 20px 24px rgba(153, 155, 168, 0.18);
+  height: 100%;
+  overflow: visible;
+  position: relative;
+
+  > div {
+    opacity: 0;
+    pointer-events: none;
+  }
+
+  ${({$isOpen}) =>
+    $isOpen &&
+    css`
+      > div {
+        opacity: 1;
+        pointer-events: auto;
+      }
+    `}
 `;
