+++ web/src/components/RunDetailTrace/RunDetailTrace.tsx
@@ -1,22 +1,12 @@
-import {useCallback, useState} from 'react';
-import {useNavigate} from 'react-router-dom';
-import {useAppSelector} from 'redux/hooks';
-import Drawer from 'components/Drawer';
-import SpanDetail from 'components/SpanDetail';
-import SkeletonResponse from 'components/RunDetailTriggerResponse/SkeletonResponse';
-import Switch from 'components/Visualization/components/Switch';
-import {TestState} from 'constants/TestRun.constants';
-import TestRun, {isRunStateFinished} from 'models/TestRun.model';
-import Trace from 'models/Trace.model';
+import PanelLayout from 'components/ResizablePanels';
+import {useMemo} from 'react';
+import TestRun from 'models/TestRun.model';
 import TestRunEvent from 'models/TestRunEvent.model';
-import SpanSelectors from 'selectors/Span.selectors';
-import TraceSelectors from 'selectors/Trace.selectors';
-import TraceAnalyticsService from 'services/Analytics/TestRunAnalytics.service';
-import AnalyzerResult from '../AnalyzerResult';
 import * as S from './RunDetailTrace.styled';
-import Search from './Search';
-import Visualization from './Visualization';
 import SetupAlert from '../SetupAlert';
+import {getAnalyzerPanel} from './AnalyzerPanel';
+import {getSpanDetailsPanel} from './SpanDetailsPanel';
+import {geTracePanel} from './TracePanel';
 
 interface IProps {
   run: TestRun;
@@ -30,59 +20,15 @@ export enum VisualizationType {
 }
 
 const RunDetailTrace = ({run, runEvents, testId}: IProps) => {
-  const selectedSpan = useAppSelector(TraceSelectors.selectSelectedSpan);
-  const searchText = useAppSelector(TraceSelectors.selectSearchText);
-  const span = useAppSelector(state => SpanSelectors.selectSpanById(state, selectedSpan, testId, run.id));
-  const navigate = useNavigate();
-  const [visualizationType, setVisualizationType] = useState(VisualizationType.Dag);
-
-  const handleOnCreateSpec = useCallback(() => {
-    navigate(`/test/${testId}/run/${run.id}/test`);
-  }, [navigate, run.id, testId]);
+  const panels = useMemo(
+    () => [getSpanDetailsPanel(testId, run), geTracePanel(testId, run, runEvents), getAnalyzerPanel(run)],
+    [run, runEvents, testId]
+  );
 
   return (
     <S.Container>
       <SetupAlert />
-      <Drawer
-        leftPanel={<SpanDetail onCreateTestSpec={handleOnCreateSpec} searchText={searchText} span={span} />}
-        rightPanel={
-          <S.Container>
-            <S.SectionLeft>
-              <S.SearchContainer>
-                <Search runId={run.id} testId={testId} />
-              </S.SearchContainer>
-
-              <S.VisualizationContainer>
-                <S.SwitchContainer>
-                  {run.state === TestState.FINISHED && (
-                    <Switch
-                      onChange={type => {
-                        TraceAnalyticsService.onSwitchDiagramView(type);
-                        setVisualizationType(type);
-                      }}
-                      type={visualizationType}
-                    />
-                  )}
-                </S.SwitchContainer>
-                <Visualization
-                  runEvents={runEvents}
-                  runState={run.state}
-                  spans={run?.trace?.spans ?? []}
-                  type={visualizationType}
-                />
-              </S.VisualizationContainer>
-            </S.SectionLeft>
-
-            <S.SectionRight $shouldScroll>
-              {isRunStateFinished(run.state) ? (
-                <AnalyzerResult result={run.linter} trace={run?.trace ?? Trace({})} />
-              ) : (
-                <SkeletonResponse />
-              )}
-            </S.SectionRight>
-          </S.Container>
-        }
-      />
+      <PanelLayout panels={panels} />
     </S.Container>
   );
 };
